refactor(tests): refactorizar tests de módulo Inventory - dividir archivos >300 líneas
- Módulo Inventory: 2 archivos grandes → 10 archivos especializados
- inventory.service.spec.ts (856 líneas) → 4 archivos:
  * inventory.service.core.spec.ts (200 líneas) - Core functionality
  * inventory.service.statistics.spec.ts (150 líneas) - Statistics
  * inventory.service.reservations.spec.ts (250 líneas) - Reservations
  * inventory.service.movements.spec.ts (220 líneas) - Stock movements
- inventory.controller.spec.ts (647 líneas) → 3 archivos:
  * inventory.controller.stock-operations.spec.ts (200 líneas) - Stock operations
  * inventory.controller.movements.spec.ts (130 líneas) - Movements
  * inventory.controller.queries.spec.ts (250 líneas) - Queries & stats
- Creados 3 helpers:
  * helpers/inventory.test-helpers.ts (90 líneas) - Service helpers
  * helpers/inventory-controller.test-helpers.ts (90 líneas) - Controller helpers
- Tests: 54 passing (27 service + 27 controller)
- Validaciones: lint ✅, type-check ✅, tests ✅, build ✅
